pyproject.toml
[project]
name = "student-data-exfiltration-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "openai-agents>=0.4.1",
    "openai-sdk>=0.1.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.1",
    "tavily-python>=0.7.12",
]
